+++ config/config_test.go
@@ -37,13 +37,13 @@ func TestNewConfig(t *testing.T) {
 			},
 			expectError: false,
 			expected: &Config{
-				GithubToken:   "gh_token123",
-				FreshRssUrl:   "http://freshrss.example.com",
-				FreshRssUser:  "testuser",
-				FreshRssToken: "freshrss_token456",
+				GitHubToken:   "gh_token123",
+				FreshRSSURL:   "http://freshrss.example.com",
+				FreshRSSUser:  "testuser",
+				FreshRSSToken: "freshrss_token456",
 				DebugMode:     false,
 				SingleRunMode: false,
-				HttpTimeout:   10 * time.Second,
+				HTTPTimeout:   10 * time.Second,
 			},
 		},
 		{
@@ -59,13 +59,13 @@ func TestNewConfig(t *testing.T) {
 			},
 			expectError: false,
 			expected: &Config{
-				GithubToken:   "gh_token123",
-				FreshRssUrl:   "http://freshrss.example.com",
-				FreshRssUser:  "testuser",
-				FreshRssToken: "freshrss_token456",
+				GitHubToken:   "gh_token123",
+				FreshRSSURL:   "http://freshrss.example.com",
+				FreshRSSUser:  "testuser",
+				FreshRSSToken: "freshrss_token456",
 				DebugMode:     true,
 				SingleRunMode: true,
-				HttpTimeout:   30 * time.Second,
+				HTTPTimeout:   30 * time.Second,
 			},
 		},
 		{
@@ -119,13 +119,13 @@ func TestNewConfig(t *testing.T) {
 			},
 			expectError: false,
 			expected: &Config{
-				GithubToken:   "gh_token123",
-				FreshRssUrl:   "http://freshrss.example.com",
-				FreshRssUser:  "testuser",
-				FreshRssToken: "freshrss_token456",
+				GitHubToken:   "gh_token123",
+				FreshRSSURL:   "http://freshrss.example.com",
+				FreshRSSUser:  "testuser",
+				FreshRSSToken: "freshrss_token456",
 				DebugMode:     false,
 				SingleRunMode: false,
-				HttpTimeout:   10 * time.Second,
+				HTTPTimeout:   10 * time.Second,
 			},
 		},
 		{
@@ -139,13 +139,13 @@ func TestNewConfig(t *testing.T) {
 			},
 			expectError: false,
 			expected: &Config{
-				GithubToken:   "gh_token123",
-				FreshRssUrl:   "http://freshrss.example.com",
-				FreshRssUser:  "testuser",
-				FreshRssToken: "freshrss_token456",
+				GitHubToken:   "gh_token123",
+				FreshRSSURL:   "http://freshrss.example.com",
+				FreshRSSUser:  "testuser",
+				FreshRSSToken: "freshrss_token456",
 				DebugMode:     false,
 				SingleRunMode: false,
-				HttpTimeout:   10 * time.Second,
+				HTTPTimeout:   10 * time.Second,
 			},
 		},
 	}
@@ -167,20 +167,20 @@ func TestNewConfig(t *testing.T) {
 				return
 			}
 
-			if config.GithubToken != tc.expected.GithubToken {
-				t.Errorf("Expected GithubToken %s, got %s", tc.expected.GithubToken, config.GithubToken)
+			if config.GitHubToken != tc.expected.GitHubToken {
+				t.Errorf("Expected GitHubToken %s, got %s", tc.expected.GitHubToken, config.GitHubToken)
 			}
 
-			if config.FreshRssUrl != tc.expected.FreshRssUrl {
-				t.Errorf("Expected FreshRssUrl %s, got %s", tc.expected.FreshRssUrl, config.FreshRssUrl)
+			if config.FreshRSSURL != tc.expected.FreshRSSURL {
+				t.Errorf("Expected FreshRSSUrl %s, got %s", tc.expected.FreshRSSURL, config.FreshRSSURL)
 			}
 
-			if config.FreshRssUser != tc.expected.FreshRssUser {
-				t.Errorf("Expected FreshRssUser %s, got %s", tc.expected.FreshRssUser, config.FreshRssUser)
+			if config.FreshRSSUser != tc.expected.FreshRSSUser {
+				t.Errorf("Expected FreshRSSUser %s, got %s", tc.expected.FreshRSSUser, config.FreshRSSUser)
 			}
 
-			if config.FreshRssToken != tc.expected.FreshRssToken {
-				t.Errorf("Expected FreshRssToken %s, got %s", tc.expected.FreshRssToken, config.FreshRssToken)
+			if config.FreshRSSToken != tc.expected.FreshRSSToken {
+				t.Errorf("Expected FreshRSSToken %s, got %s", tc.expected.FreshRSSToken, config.FreshRSSToken)
 			}
 
 			if config.DebugMode != tc.expected.DebugMode {
@@ -191,8 +191,8 @@ func TestNewConfig(t *testing.T) {
 				t.Errorf("Expected SingleRunMode %t, got %t", tc.expected.SingleRunMode, config.SingleRunMode)
 			}
 
-			if config.HttpTimeout != tc.expected.HttpTimeout {
-				t.Errorf("Expected HttpTimeout %v, got %v", tc.expected.HttpTimeout, config.HttpTimeout)
+			if config.HTTPTimeout != tc.expected.HTTPTimeout {
+				t.Errorf("Expected HTTPTimeout %v, got %v", tc.expected.HTTPTimeout, config.HTTPTimeout)
 			}
 		})
 	}